+++ src/components/Header/DropDown/DropDown.tsx
@@ -1,17 +1,15 @@
-import {FC, ReactElement, useRef} from "react";
+import {FC, ReactElement,useRef} from "react";
 import styles from './DropDown.module.sass';
 import ProfileDropDown from "/src/components/Header/DropDown/ProfileDropDown/ProfileDropDown";
 import NotificationDropDown from "/src/components/Header/DropDown/NotificationDropdown/NotificationDropDown";
 import EntertainmentGenresDropDown
     from "/src/components/Header/DropDown/EntertainmentGenresDropDown/EntertainmentGenresDropDown";
 import TvDropDown from "/src/components/Header/DropDown/TvDropDown/TvDropDown";
-import {DropDownType} from "/src/components/Header/Header";
-import {CSSTransition} from "react-transition-group";
+import {DropDownType} from "/src/components/Header/Header.utils";
 
 interface DropDownProps {
     dropDownType: DropDownType
 }
-
 
 const DropDown: FC<DropDownProps> = ({dropDownType}) => {
     const dropDownContentRef = useRef<HTMLDivElement>(null);
@@ -32,16 +30,9 @@ const DropDown: FC<DropDownProps> = ({dropDownType}) => {
     }
 
     return (
-    <CSSTransition classNames={{
-        enter: styles["dropdown-enter"],
-        enterActive: styles["dropdown-enter-active"],
-        exit: styles["dropdown-exit"],
-        exitActive: styles["dropdown-exit-active"],
-    }} nodeRef={dropDownContentRef} timeout={400} unmountOnExit>
         <div ref={dropDownContentRef} className={styles.dropdown__container}>
-            { getDropDownContent(dropDownType) }
+            {getDropDownContent(dropDownType)}
         </div>
-    </CSSTransition>
     )
 }
