+++ src/components/Header/DropDown/ProfileDropDown/ProfileDropDown.module.sass
@@ -42,7 +42,8 @@
     padding-left: 10px
     flex-grow: 1
     border-left: 1px solid rgba(255, 255, 255, .16)
-
+    button
+      margin-right: 10px
     &_links
       display: flex
       flex-direction: column
