+++ src/components/Header/DropDown/TvDropDown/TvDropDown.module.sass
@@ -5,6 +5,7 @@
     flex-direction: column
     justify-content: space-between
     border-right: 1px solid rgba(255,255,255,.16)
+    padding-right: 10px
     .tv
       &__links
         &__title
